+++ core/CHANGELOG.md
@@ -3,11 +3,11 @@
 - Introduce `StreamMuxerEvent::map_inbound_stream`. See [PR 2691].
 - Remove `{read,write,flush,shutdown,destroy}_substream` functions from `StreamMuxer` trait
   in favor of forcing `StreamMuxer::Substream` to implement `AsyncRead + AsyncWrite`. See [PR 2707].
-- Remove `Into<std::io::Error>` bound from `StreamMuxer::Error`. See [PR XXXX].
+- Remove `Into<std::io::Error>` bound from `StreamMuxer::Error`. See [PR 2710].
 
 [PR 2691]: https://github.com/libp2p/rust-libp2p/pull/2691
 [PR 2707]: https://github.com/libp2p/rust-libp2p/pull/2707
-[PR XXXX]: https://github.com/libp2p/rust-libp2p/pull/XXXX
+[PR 2710]: https://github.com/libp2p/rust-libp2p/pull/2710
 
 # 0.33.0